// src/api/models/addLinkDto.ts
/**
 * Generated by orval v6.10.2 🍺
 * Do not edit manually.
 * API
 * API DOCS
 * OpenAPI spec version: 1.0
 */
import type { AddLinkDtoLabels } from './addLinkDtoLabels';

export interface AddLinkDto {
  appId: string;
  value: string;
  type: string;
  categoryId: string;
  labels?: AddLinkDtoLabels;
  icon?: boolean;
  enabled?: boolean;
}