// src/api/models/editLinkDto.ts
/**
 * Generated by orval v6.10.2 🍺
 * Do not edit manually.
 * API
 * API DOCS
 * OpenAPI spec version: 1.0
 */
import type { EditLinkDtoLabels } from './editLinkDtoLabels';

export interface EditLinkDto {
  appId: string;
  categoryId: string;
  value?: string;
  icon?: boolean;
  enabled?: boolean;
  labels?: EditLinkDtoLabels;
}